src/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
Django==4.1.3
django-elasticsearch-dsl==7.2.2
django-isbn-field==0.5.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
flake8==5.0.4
mccabe==0.7.0
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-stdnum==1.17
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
